PortafolioBryam/requirements.txt
pillow
django
django-admin-interface
channels==2.4.0
channels_redis==2.4.2
django-embed-video==1.3.3
twisted
websocket-client

